requirements.txt
python-dotenv
alpaca-trade-api
pandas
numpy
scikit-learn
xgboost
matplotlib
plotly
pyarrow
joblib

